move core query functions to repository files, import these repositories in core file
